pyproject.toml
[tool.poetry]
name = "star-wars-backend"
version = "0.1.0"
description = ""
authors = ["Василий Субботин <subv13@yandex.ru>"]
packages = [
    { include = "src", from = "." }
]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.70.0"
Jinja2 = "^3.0.2"
uvicorn = { extras = ["standard"], version = "^0.15.0" }
aiohttp = "^3.7.4"
attrs = "^21.2.0"
strawberry-graphql = "^0.84.1"
aioredis = "^2.0.0"
databases = { extras = ["postgresql"], version = "^0.5.3" }
SQLAlchemy = "^1.4.26"
psycopg2-binary = "^2.9.1"
dependency-injector = "^4.37.0"
typer = "^0.4.0"
pydantic = "^1.8.2"

[tool.poetry.scripts]
cli = "src.cli.main:cli"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-aiohttp = "^0.3.0"
pre-commit = "^2.15.0"
mypy = "^0.910"
black = "^21.9b0"
isort = "^5.9.3"
ipython = "^7.28.0"
types-attrs = "^19.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119

[tool.isort]
line_length = 119
combine_as_imports = true
balanced_wrapping = true
profile = "black"
multi_line_output = 3
include_trailing_comma = true
